:fix: / feat/user / filecontoller error

See merge request s03-blockchain-sub2/s03p22a104!22

+++ backend/src/main/java/com/project/model/user/UserEntity.java
@@ -21,7 +21,7 @@ public class UserEntity {
 	@Id
 	@Column(name = "uid")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	int uid; // uid
+	long uid; // uid
 	
 	@Column(name = "upw")
     String upw; // passwd
